pyproject.toml
[project]
name = "discord-gemini-bot"
version = "0.1.0"
description = "discord gemini bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.8.0",
    "nextcord>=3.0.1",
    "setuptools>=78.1.0",
]
